+++ BarHide/ViewControllers/GeneralVC.swift
@@ -25,8 +25,8 @@ final class General: NSViewController, SettingsPane {
     @IBOutlet private var HideStatusBarIconsAtLaunchCheckbox: NSButton!
     @IBOutlet private var HideStatusBarIconsAfterDelayCheckbox: NSButton!
     @IBOutlet private var HideStatusBarIconsSecondsPopUpButton: NSPopUpButton!
-    @IBOutlet private var HideBothDozerIconsCheckbox: NSButton!
-    @IBOutlet private var EnableRemoveDozerIconCheckbox: NSButton!
+    @IBOutlet private var HideBothBarHideIconsCheckbox: NSButton!
+    @IBOutlet private var EnableRemoveBarHideIconCheckbox: NSButton!
     @IBOutlet private var ShowIconAndMenuCheckbox: NSButton!
     @IBOutlet private var FontSizePopUpButton: NSPopUpButton!
     @IBOutlet private var ButtonPaddingPopUpButton: NSPopUpButton!
@@ -48,8 +48,8 @@ final class General: NSViewController, SettingsPane {
 
         HideStatusBarIconsAtLaunchCheckbox.isChecked = Defaults[.hideAtLaunchEnabled]
         HideStatusBarIconsAfterDelayCheckbox.isChecked = Defaults[.hideAfterDelayEnabled]
-        HideBothDozerIconsCheckbox.isChecked = Defaults[.noIconMode]
-        EnableRemoveDozerIconCheckbox.isChecked = Defaults[.removeDozerIconEnabled]
+        HideBothBarHideIconsCheckbox.isChecked = Defaults[.noIconMode]
+        EnableRemoveBarHideIconCheckbox.isChecked = Defaults[.removeBarHideIconEnabled]
         ShowIconAndMenuCheckbox.isChecked = Defaults[.showIconAndMenuEnabled]
         HideStatusBarIconsSecondsPopUpButton.selectItem(withTitle: "\(Int(Defaults[.hideAfterDelay])) seconds")
         FontSizePopUpButton.selectItem(withTitle: "\(Int(Defaults[.iconSize])) px")
@@ -79,45 +79,45 @@ final class General: NSViewController, SettingsPane {
     }
 
     @IBAction private func hideStatusBarIconsAtLaunchClicked(_ sender: NSButton) {
-        DozerIcons.shared.hideStatusBarIconsAtLaunch = HideStatusBarIconsAtLaunchCheckbox.isChecked
+        BarHideIcons.shared.hideStatusBarIconsAtLaunch = HideStatusBarIconsAtLaunchCheckbox.isChecked
     }
 
     @IBAction private func hideStatusBarIconsAfterDelayClicked(_ sender: NSButton) {
-        DozerIcons.shared.hideStatusBarIconsAfterDelay = HideStatusBarIconsAfterDelayCheckbox.isChecked
+        BarHideIcons.shared.hideStatusBarIconsAfterDelay = HideStatusBarIconsAfterDelayCheckbox.isChecked
     }
 
     @IBAction private func hideStatusBarIconsSecondsUpdated(_ sender: NSPopUpButton) {
         Defaults[.hideAfterDelay] = TimeInterval(HideStatusBarIconsSecondsPopUpButton.selectedTag())
-        DozerIcons.shared.resetTimer()
+        BarHideIcons.shared.resetTimer()
     }
 
-    @IBAction private func hideBothDozerIconsClicked(_ sender: NSButton) {
-        DozerIcons.shared.hideBothDozerIcons = HideBothDozerIconsCheckbox.isChecked
+    @IBAction private func hideBothBarHideIconsClicked(_ sender: NSButton) {
+        BarHideIcons.shared.hideBothBarHideIcons = HideBothBarHideIconsCheckbox.isChecked
     }
 
     @IBAction private func showIconAndMenuClicked(_ sender: NSButton) {
-        DozerIcons.shared.enableIconAndMenu = ShowIconAndMenuCheckbox.isChecked
+        BarHideIcons.shared.enableIconAndMenu = ShowIconAndMenuCheckbox.isChecked
     }
 
     @IBAction private func fontSizeChanged(_ sender: NSPopUpButton) {
-        DozerIcons.shared.iconFontSize = FontSizePopUpButton.selectedTag()
+        BarHideIcons.shared.iconFontSize = FontSizePopUpButton.selectedTag()
     }
 
     @IBAction private func buttonPaddingChanged(_ sender: NSPopUpButton) {
-        DozerIcons.shared.buttonPadding = CGFloat(ButtonPaddingPopUpButton.selectedTag())
+        BarHideIcons.shared.buttonPadding = CGFloat(ButtonPaddingPopUpButton.selectedTag())
     }
 
-    @IBAction private func enableRemoveDozerIconClicked(_ sender: NSButton) {
-        DozerIcons.shared.enableRemoveDozerIcon = EnableRemoveDozerIconCheckbox.isChecked
+    @IBAction private func enableRemoveBarHideIconClicked(_ sender: NSButton) {
+        BarHideIcons.shared.enableRemoveBarHideIcon = EnableRemoveBarHideIconCheckbox.isChecked
     }
 
     /// disables the noIcon-checkbox if no shortcut is set and keeps track whether shortcut is set
     private func configureEnabledNoIconCheckbox() {
         if KeyboardShortcuts.getShortcut(for: .toggleMenuItems) == nil {
-            HideBothDozerIconsCheckbox.isEnabled = false
+            HideBothBarHideIconsCheckbox.isEnabled = false
             Defaults[.isShortcutSet] = false
         } else {
-            HideBothDozerIconsCheckbox.isEnabled = true
+            HideBothBarHideIconsCheckbox.isEnabled = true
             Defaults[.isShortcutSet] = true
         }
     }